# docs/swagger.yaml
definitions:
...
  group.DissolveParam:
    properties:
      group_id:
        type: string
    required:
    - group_id
    type: object
  group.HandleJoinRequestParam:
    properties:
      action:
        description: approved | rejected
        type: string
      request_id:
        type: string
    required:
    - action
    - request_id
    type: object
...
  group.JoinOrLeaveParam:
    properties:
      group_id:
        type: string
    required:
    - group_id
    type: object
...
  group.MarkReadParam:
    properties:
      group_id:
        type: string
      message_id:
        type: string
    required:
    - group_id
    - message_id
    type: object
...
  group.RecallMessageParam:
    properties:
      group_id:
        type: string
      message_id:
        type: string
    required:
    - group_id
    - message_id
    type: object
  group.SendMessageParam:
    properties:
      content:
        type: string
      extra:
        description: JSON TEXT
        type: string
      group_id:
        type: string
      msg_type:
        type: string
      reply_to:
        type: string
    required:
    - group_id
    type: object
...
info:
  contact:
    name: Charlie
    url: https://github.com/jiangbyte
  description: Hei Gin 后台管理系统 API 文档
  license:
    name: MIT
    url: https://github.com/jiangbyte/hei-gin/blob/main/LICENSE
  termsOfService: https://github.com/jiangbyte/hei-gin
  title: Hei Gin API
  version: 1.0.0
paths:
...
  /api/v1/c/im/group/join:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/c/im/group/join，即时通讯群组加入群组
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.JoinOrLeaveParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组加入群组
      tags:
      - 即时通讯群组
  /api/v1/c/im/group/leave:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/c/im/group/leave，即时通讯群组退出群组
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.JoinOrLeaveParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组退出群组
      tags:
      - 即时通讯群组
  /api/v1/c/im/group/mark-read:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/mark-read，即时通讯群组标记已读
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.MarkReadParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组标记已读
      tags:
      - 即时通讯群组
...
  /api/v1/c/im/group/messages:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/messages，即时通讯群组消息列表
      parameters:
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: cursor
        in: query
        name: cursor
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组消息列表
      tags:
      - 即时通讯群组
...
  /api/v1/c/im/group/recall:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/recall，即时通讯群组撤回消息
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.RecallMessageParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组撤回消息
      tags:
      - 即时通讯群组
  /api/v1/c/im/group/search:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/search，即时通讯群组搜索
      parameters:
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: cursor
        in: query
        name: cursor
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组搜索
      tags:
      - 即时通讯群组
  /api/v1/c/im/group/search-groups:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/search-groups，即时通讯群组群组搜索
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组群组搜索
      tags:
      - 即时通讯群组
  /api/v1/c/im/group/send:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/send，即时通讯群组发送消息
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.SendMessageParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组发送消息
      tags:
      - 即时通讯群组
...
  /api/v1/sys/im/group/dissolve:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/dissolve，即时通讯群组解散群组
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.DissolveParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组解散群组
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/handle-join-request:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/handle-join-request，即时通讯群组处理入群申请
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.HandleJoinRequestParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组处理入群申请
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/invite:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/invite，即时通讯群组邀请成员
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.InviteParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组邀请成员
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/join:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/join，即时通讯群组加入群组
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.JoinOrLeaveParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组加入群组
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/kick:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/kick，即时通讯群组移除成员
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.KickParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组移除成员
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/leave:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/leave，即时通讯群组退出群组
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.JoinOrLeaveParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组退出群组
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/mark-read:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/mark-read，即时通讯群组标记已读
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.MarkReadParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组标记已读
      tags:
      - 即时通讯群组
...
  /api/v1/sys/im/group/messages:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/messages，即时通讯群组消息列表
      parameters:
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: cursor
        in: query
        name: cursor
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组消息列表
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/mute:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/mute，即时通讯群组禁言
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.MuteParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组禁言
      tags:
      - 即时通讯群组
...
  /api/v1/sys/im/group/recall:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/recall，即时通讯群组撤回消息
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.RecallMessageParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组撤回消息
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/search:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/search，即时通讯群组搜索
      parameters:
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: cursor
        in: query
        name: cursor
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组搜索
      tags:
      - 即时通讯群组
  /api/v1/sys/im/group/search-groups:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/search-groups，即时通讯群组群组搜索
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组群组搜索
      tags:
      - 即时通讯群组
...
  /api/v1/sys/im/group/update:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/group/update，即时通讯群组更新
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/group.UpdateParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯群组更新
      tags:
      - 即时通讯群组
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"